+++ App/controllers/admin.py
@@ -1,4 +1,4 @@
-from App.models import Admin
+from App.models import Admin, ClashRule
 from App.database import db
 
 def login_admin(email, password):
@@ -7,3 +7,16 @@ def login_admin(email, password):
         if admin.check_password(password):
             return admin.login()
     return "Login failed"
+
+def custom_clash_rule(clashRuleID, userID, clashRuleTitle, clashRuleDescription, allowableDays): #create_new_clash_rule? as a possible function name
+    rule = ClashRule.query.get(clashRuleID)
+    if rule: 
+        return rule
+    else:
+         #Add new rule
+        newRule = ClashRule.addNewRule(clashRuleID, userID, clashRuleTitle, clashRuleDescription, allowableDays) #possible name "addNewRule?" for the adding new rules/custome rules that would be in the models section
+        db.session.add(newRule)
+        db.session.commit()
+        return newRule
+
+    #Note the class and function call names and the ordering of the parameter might have to change depending on the models
